--- hudi-common/src/main/java/org/apache/hudi/common/util/MapUtils.java
@@ -1,41 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package org.apache.hudi.common.util;
-
-import java.util.Map;
-import java.util.Objects;
-
-/**
- * Utils for Java Map.
- */
-public class MapUtils {
-
-  public static boolean isNullOrEmpty(Map<?, ?> m) {
-    return Objects.isNull(m) || m.isEmpty();
-  }
-
-  public static boolean nonEmpty(Map<?, ?> m) {
-    return !isNullOrEmpty(m);
-  }
-
-  public static boolean containsAll(Map<?, ?> m1, Map<?, ?> m2) {
-    return m1.entrySet().containsAll(m2.entrySet());
-  }
-}
